Dispute detail: Add dispute final outcome and round number to info section
