+++ src/components/Disputes/DisputeInfo.js
@@ -7,15 +7,24 @@ import {
   textStyle,
   TransactionBadge,
   useTheme,
+  isAddress,
 } from '@aragon/ui'
+import styled from 'styled-components'
 
 import DisputeStatus from './DisputeStatus'
 import DisputeCurrentRuling from './DisputeCurrentRuling'
 import DisputeActions from './DisputeActions'
 import Loading from './Loading'
+import { useConnectedAccount } from '../../providers/Web3'
 
-import { Phase as DisputePhase } from '../../types/dispute-status-types'
+import {
+  Phase as DisputePhase,
+  Status as DipsuteStatus,
+} from '../../types/dispute-status-types'
 import iconCourt from '../../assets/courtIcon.svg'
+import { addressesEqual } from '../../lib/web3-utils'
+import DisputeRoundPill from './DisputeRoundPill'
+import DisputeOutcomeText from './DisputeOutcomeText'
 
 const DisputeInfo = React.memo(function({
   id,
@@ -28,12 +37,13 @@ const DisputeInfo = React.memo(function({
   onRequestAppeal,
   onExecuteRuling,
 }) {
-  const theme = useTheme()
-  const { phase } = dispute || {}
+  const { phase, status } = dispute || {}
 
   const description = dispute && dispute.metadata
   const creatorAddress = dispute && dispute.subject && dispute.subject.id
-  const transaction = dispute && dispute.txHash
+
+  const isFinalRulingEnsured =
+    phase === DisputePhase.ExecuteRuling || status === DipsuteStatus.Closed
 
   return (
     <Box>
@@ -45,121 +55,44 @@ const DisputeInfo = React.memo(function({
           align-items: center;
         `}
       >
-        <div
-          css={`
-            display: flex;
-            margin-bottom: ${3 * GU}px;
-          `}
-        >
-          <div
-            css={`
-              display: flex;
-              align-items: center;
-            `}
-          >
-            <div
-              css={`
-                background: linear-gradient(
-                  233deg,
-                  ${theme.accentEnd} -50%,
-                  ${theme.accentStart} 91%
-                );
-                border-radius: 50%;
-                padding: 12px;
-                display: inline-block;
-              `}
-            >
-              <img src={iconCourt} alt="" width="39" height="32" />
-            </div>
-            <div
-              css={`
-                margin-left: ${3 * GU}px;
-              `}
-            >
-              <h1
-                css={`
-                  display: flex;
-                  align-items: center;
-                  margin-bottom: ${1 * GU}px;
-                  ${textStyle('title3')};
-                `}
-              >
-                <span>Dispute #{id}</span>
-                {dispute && (
-                  <DisputeStatus
-                    dispute={dispute}
-                    css={`
-                      margin: 0 0 0 ${1 * GU}px;
-                    `}
-                  />
-                )}
-              </h1>
-              {transaction && <TransactionBadge transaction={transaction} />}
-            </div>
-          </div>
-        </div>
+        <DisputeHeader id={id} dispute={dispute} />
         {loading ? (
           <Loading border={false} />
         ) : (
-          <div
-            css={`
-              display: grid;
-              grid-template-columns: 1fr minmax(250px, auto);
-              grid-gap: ${5 * GU}px;
-              margin-bottom: ${2 * GU}px;
-            `}
-          >
-            <div>
-              <h2
-                css={`
-                  ${textStyle('label2')};
-                  color: ${theme.surfaceContentSecondary};
-                  margin-bottom: ${2 * GU}px;
-                `}
-              >
-                Description
-              </h2>
-              <span
-                css={`
-                  ${textStyle('body2')};
-                `}
-              >
-                {description}
-              </span>
-            </div>
-            {creatorAddress && (
-              <div>
-                <h2
-                  css={`
-                    ${textStyle('label2')};
-                    color: ${theme.surfaceContentSecondary};
-                    margin-bottom: ${2 * GU}px;
-                  `}
-                >
-                  Created by
-                </h2>
-                <div
-                  css={`
-                    display: flex;
-                    align-items: flex-start;
-                  `}
-                >
-                  <IdentityBadge
-                    // connectedAccount={addressesEqual(creator, connectedAccount)} TODO- add connected account
-                    entity={creatorAddress}
-                  />
-                </div>
-              </div>
+          <>
+            {isFinalRulingEnsured && (
+              <Row>
+                <Field
+                  label="Final jury outcome"
+                  value={
+                    <DisputeOutcomeText
+                      outcome={
+                        dispute.rounds[dispute.lastRoundId].vote.winningOutcome
+                      }
+                      phase={dispute.phase}
+                      disputeEnded={isFinalRulingEnsured}
+                    />
+                  }
+                />
+                <Field
+                  label="Round number"
+                  value={<DisputeRoundPill roundId={dispute.lastRoundId} />}
+                  isNode
+                />
+              </Row>
             )}
-          </div>
+            <Row>
+              <Field label="Description" value={description} />
+              {creatorAddress && (
+                <Field label="Plantiff" value={creatorAddress} />
+              )}
+            </Row>
+          </>
         )}
 
         {(phase === DisputePhase.AppealRuling ||
           phase === DisputePhase.ConfirmAppeal ||
-          phase === DisputePhase.ExecuteRuling ||
-          phase === DisputePhase.ClaimRewards) && (
-          <DisputeCurrentRuling dispute={dispute} />
-        )}
+          isFinalRulingEnsured) && <DisputeCurrentRuling dispute={dispute} />}
         {!loading && (
           <DisputeActions
             dispute={dispute}
@@ -175,6 +108,122 @@ const DisputeInfo = React.memo(function({
     </Box>
   )
 })
+
+function DisputeHeader({ id, dispute }) {
+  const theme = useTheme()
+  const transaction = dispute && dispute.txHash
+
+  return (
+    <div
+      css={`
+        display: flex;
+        margin-bottom: ${3 * GU}px;
+      `}
+    >
+      <div
+        css={`
+          display: flex;
+          align-items: center;
+        `}
+      >
+        <div
+          css={`
+            background: linear-gradient(
+              233deg,
+              ${theme.accentEnd} -50%,
+              ${theme.accentStart} 91%
+            );
+            border-radius: 50%;
+            padding: 12px;
+            display: inline-block;
+          `}
+        >
+          <img src={iconCourt} alt="" width="39" height="32" />
+        </div>
+        <div
+          css={`
+            margin-left: ${3 * GU}px;
+          `}
+        >
+          <h1
+            css={`
+              display: flex;
+              align-items: center;
+              margin-bottom: ${1 * GU}px;
+              ${textStyle('title3')};
+            `}
+          >
+            <span>Dispute #{id}</span>
+            {dispute && (
+              <DisputeStatus
+                dispute={dispute}
+                css={`
+                  margin: 0 0 0 ${1 * GU}px;
+                `}
+              />
+            )}
+          </h1>
+          {transaction && <TransactionBadge transaction={transaction} />}
+        </div>
+      </div>
+    </div>
+  )
+}
+
+function Field({ label, value, isNode }) {
+  const theme = useTheme()
+  const connectedAccount = useConnectedAccount()
+
+  return (
+    <div>
+      <h2
+        css={`
+          ${textStyle('label2')};
+          color: ${theme.surfaceContentSecondary};
+          margin-bottom: ${1 * GU}px;
+        `}
+      >
+        {label}
+      </h2>
+
+      {(() => {
+        if (isAddress(value))
+          return (
+            <div
+              css={`
+                display: flex;
+                align-items: flex-start;
+              `}
+            >
+              <IdentityBadge
+                connectedAccount={addressesEqual(value, connectedAccount)}
+                entity={value}
+              />
+            </div>
+          )
+
+        if (isNode) return <>{value}</>
+
+        return (
+          <span
+            css={`
+              ${textStyle('body2')};
+            `}
+          >
+            {value}
+          </span>
+        )
+      })()}
+    </div>
+  )
+}
+
+const Row = styled.div`
+  display: grid;
+  grid-template-columns: 1fr minmax(250px, auto);
+  grid-gap: ${5 * GU}px;
+  margin-bottom: ${2 * GU}px;
+`
 
 DisputeInfo.propTypes = {
   dispute: PropTypes.object, // TODO: define DisputeType
